iceberg: pull out reusable bits of manifest_io
Pulls out some generic-looking code from manifest_io. I intend on using
this for other kinds of metadata, and want to separate out IO based on
how they're used (manifests vs table metadata).
