+++ src/v/iceberg/manifest_io.h
@@ -9,18 +9,16 @@
 #pragma once
 
 #include "base/outcome.h"
-#include "bytes/iobuf.h"
 #include "cloud_io/remote.h"
+#include "cloud_storage_clients/types.h"
 #include "iceberg/manifest.h"
 #include "iceberg/manifest_list.h"
+#include "iceberg/metadata_io.h"
 #include "iceberg/partition_key_type.h"
-#include "model/fundamental.h"
 #include "utils/named_type.h"
 
 #include <seastar/core/future.hh>
 #include <seastar/util/noncopyable_function.hh>
-
-#include <system_error>
 
 namespace iceberg {
 
@@ -29,50 +27,22 @@ using manifest_path
 using manifest_list_path
   = named_type<std::filesystem::path, struct manifest_path_tag>;
 
-class manifest_io {
+class manifest_io : public metadata_io {
 public:
-    enum class errc {
-        // The call has timed out. Callers should retry, if allowed.
-        timedout = 1,
-
-        // The call failed for some reason that doesn't appear to be retriable.
-        failed,
-
-        // The call failed because the system is shutting down.
-        shutting_down,
-    };
     explicit manifest_io(
       cloud_io::remote& io, cloud_storage_clients::bucket_name b)
-      : io_(io)
-      , bucket_(b) {}
+      : metadata_io(io, std::move(b)) {}
     ~manifest_io() = default;
 
-    ss::future<checked<manifest, errc>> download_manifest(
+    ss::future<checked<manifest, metadata_io::errc>> download_manifest(
       const manifest_path& path, const partition_key_type& pk_type);
-    ss::future<checked<manifest_list, errc>>
+    ss::future<checked<manifest_list, metadata_io::errc>>
     download_manifest_list(const manifest_list_path& path);
 
-    ss::future<checked<size_t, errc>>
+    ss::future<checked<size_t, metadata_io::errc>>
     upload_manifest(const manifest_path& path, const manifest&);
-    ss::future<checked<size_t, errc>>
+    ss::future<checked<size_t, metadata_io::errc>>
     upload_manifest_list(const manifest_list_path& path, const manifest_list&);
-
-private:
-    template<typename T>
-    ss::future<checked<T, errc>> download_object(
-      const std::filesystem::path& path,
-      std::string_view display_str,
-      ss::noncopyable_function<T(iobuf)>);
-
-    template<typename T>
-    ss::future<checked<size_t, manifest_io::errc>> upload_object(
-      const std::filesystem::path& path,
-      const T&,
-      std::string_view display_str,
-      ss::noncopyable_function<iobuf(const T&)>);
-
-    cloud_io::remote& io_;
-    const cloud_storage_clients::bucket_name bucket_;
 };
 
 } // namespace iceberg
