+++ src/v/iceberg/tests/manifest_io_test.cc
@@ -69,7 +69,7 @@ TEST_F(ManifestIOTest, TestManifestRoundtrip) {
     auto test_path = manifest_path{"foo/bar/baz"};
     auto dl_res = io.download_manifest(test_path, empty_pk_type()).get();
     ASSERT_TRUE(dl_res.has_error());
-    ASSERT_EQ(dl_res.error(), manifest_io::errc::failed);
+    ASSERT_EQ(dl_res.error(), metadata_io::errc::failed);
 
     // Now upload, then try again with success.
     auto ul_res = io.upload_manifest(test_path, m).get();
@@ -106,7 +106,7 @@ TEST_F(ManifestIOTest, TestManifestListRoundtrip) {
     auto test_path = manifest_path{"foo/bar/baz"};
     auto dl_res = io.download_manifest_list(test_path).get();
     ASSERT_TRUE(dl_res.has_error());
-    ASSERT_EQ(dl_res.error(), manifest_io::errc::failed);
+    ASSERT_EQ(dl_res.error(), metadata_io::errc::failed);
 
     // Now upload, then try again with success.
     auto ul_err = io.upload_manifest_list(test_path, m).get();
@@ -125,20 +125,20 @@ TEST_F(ManifestIOTest, TestShutdown) {
     {
         auto dl_res = io.download_manifest(test_path, empty_pk_type()).get();
         ASSERT_TRUE(dl_res.has_error());
-        ASSERT_EQ(dl_res.error(), manifest_io::errc::shutting_down);
+        ASSERT_EQ(dl_res.error(), metadata_io::errc::shutting_down);
 
         auto ul_err = io.upload_manifest(test_path, manifest{}).get();
         ASSERT_TRUE(ul_err.has_error());
-        ASSERT_EQ(ul_err.error(), manifest_io::errc::shutting_down);
+        ASSERT_EQ(ul_err.error(), metadata_io::errc::shutting_down);
     }
     {
         auto dl_res = io.download_manifest_list(test_path).get();
         ASSERT_TRUE(dl_res.has_error());
-        ASSERT_EQ(dl_res.error(), manifest_io::errc::shutting_down);
+        ASSERT_EQ(dl_res.error(), metadata_io::errc::shutting_down);
 
         auto ul_err = io.upload_manifest_list(test_path, manifest_list{}).get();
         ASSERT_TRUE(ul_err.has_error());
-        ASSERT_EQ(ul_err.error(), manifest_io::errc::shutting_down);
+        ASSERT_EQ(ul_err.error(), metadata_io::errc::shutting_down);
     }
 }
 
@@ -160,11 +160,11 @@ TEST_F(ManifestIOTest, TestCorruptedDownload) {
     {
         auto dl_res = io.download_manifest(test_path, empty_pk_type()).get();
         ASSERT_TRUE(dl_res.has_error());
-        ASSERT_EQ(dl_res.error(), manifest_io::errc::failed);
+        ASSERT_EQ(dl_res.error(), metadata_io::errc::failed);
     }
     {
         auto dl_res = io.download_manifest_list(test_path).get();
         ASSERT_TRUE(dl_res.has_error());
-        ASSERT_EQ(dl_res.error(), manifest_io::errc::failed);
+        ASSERT_EQ(dl_res.error(), metadata_io::errc::failed);
     }
 }
